Finish adding PHPUnit 10 support and test with PHP 8.2 and 8.3
=== Issues with PHPUnit 9.6 on PHP 7.4 ===

```
1) OckCyp\CoversValidator\Tests\Command\ValidateCommandTest::testPrintsConfigFileUsed
RuntimeException: Could not find PHPUnit TestSuiteMapper class
```

Restore the PHPUnit 9.3 and 9.5 logic from the master branch.

In PHPUnit 10, TestSuiteMapper was renamed to the same name as 9.3,
but takes different args. This means the map() call needs to become
part of the conditional. It also requires the ConfigurationRegistry
singleton to be populated, because otherwise:

> $ vendor/bin/covers-validator
> CoversValidator 1.6.0
> PHP Fatal error:  Uncaught AssertionError: assert(self::$instance instanceof Configuration)
>   in /vendor/phpunit/phpunit/src/TextUI/Configuration/Registry.php:93
> Stack trace:
> /vendor/phpunit/phpunit/src/TextUI/Configuration/Registry.php(93)
> /vendor/phpunit/phpunit/src/Framework/TestBuilder.php(128): Registry::get
> /vendor/phpunit/phpunit/src/Framework/TestBuilder.php(63): TestBuilder->configureTestCase
> /vendor/phpunit/phpunit/src/Framework/TestSuite.php(511): …
> /vendor/phpunit/phpunit/src/TextUI/Configuration/Xml/TestSuiteMapper.php(102): …
> /vendor/ockcyp/covers-validator/src/Model/TestCollection.php(50): TestSuiteMapper->map

=== Issues with PHPUnit 10 on PHP 8 ===

```
…\InputHandlerTest::testHandlesInputWithNoBootstrap
Error: Call to undefined method …\InputHandlerTest::prophesize()
```

https://github.com/sebastianbergmann/phpunit/blob/9.6.26/DEPRECATIONS.md#test-double-api-1
https://phpunit.de/announcements/phpunit-10.html
https://github.com/sebastianbergmann/phpunit/blob/10.0.0/ChangeLog-10.0.md
https://github.com/sebastianbergmann/phpunit/issues/4142

```
…\ConfigLoaderTest::testLoadsConfig
Error: Call to undefined method …\ConfigLoaderTest::preTestRun()
```

Restore previous logic for now, which seemed to work fine.

```
…\TestSuiteLoaderTest::testLoadsSuitePHPUnit9
Error: Class "PHPUnit\TextUI\Configuration\Loader" not found
```

Condition checked an unused PHPUnit9 classes, which happens to exist
on PHPUnit10 as well. The class that is actually used (Loader) was
renamed again.

Update the condition, and write a new test for PHPUnit 10. Note that
TestCase::getName() was removed as well, so update that to use
TestName::name() instead, same as in ValidateCommand already.

In addition, switch the condition from a class check to a version
check because otherwise:

> PHPUnit 9.6.26
> TestSuiteLoaderTest::testLoadsSuitePHPUnit10
> Error: Call to undefined method OneTestCoveringExistingTwoClassTest::name()

... because the PHPUnit 10+ class was first introduced in PHPUnit 9.3.

I realized the same logic exists in ConfigLoader::loadConfig, which
seems worth re-using (and testing) instead, so I've simplified the
test to use that instead.

```
1) OckCyp\CoversValidator\Tests\Loader\FileLoaderTest::testLoadsFile
Failed asserting that false is true.

2) OckCyp\CoversValidator\Tests\Loader\FileLoaderTest::testLoadsFile2
Failed asserting that false is true.
```

In PHPUnit 10, the FileLoader::checkAndLoad utility was removed,
and with it support for setting global variables from there.
The equivalent is simply include_once and that is what PHPUnit 10+
does as well.

The example test here, however, expected PHPUnit 9-specific behavior.
Skip it under PHPUnit 10+ and write a new test that defines a function
instead, which does not rely on emulating global scope for vars.

```
1) …\OneTestWithProviderCoveringNonExistentClassTest::testDummyTest
Data Provider method provideDummyTest() is not static

2) …\OneTestWithProviderCoveringExistingClassTest::testDummyTest
Data Provider method provideDummyTest() is not static
```

Fixed.

=== Issues with PHPUnit 10 coverage build ===

```
Error: No filter is configured, code coverage will not be processed
```

Migrate phpunit.xml.dist, and set the new one for the coverage
build only since it requires PHPUnit 10. It only runs on one
version anyway.

```
Run vendor/bin/php-coveralls
COVERALLS_REPO_TOKEN=""
Error: Process completed with exit code 1.
```

Skip upload step on forks where secret tokens don't apply.
Keep the rest of the coverage build enabled, so that it can
be tested.

Ref https://github.com/oradwell/covers-validator/issues/44.