tests: Run ValidateCommandTest cases as sub processes
This would have caught the PHPUnit 10 breakage due to
an unset ConfigurationRegistry singleton. This was missed in the
integration, because it invoked ValidateCommand in the same process
as part of an already-configured `phpunit` program, and thus the
global state is already fine.

If you comment out these three lines in src/Mode/TestCollection.php,
the previous tests would pass, but the new tests fail:

```php
 \PHPUnit\TextUI\Configuration\Registry::init(
    (new \PHPUnit\TextUI\CliArguments\Builder())->fromParameters([]),
    $configuration
);
```

Follows-up https://github.com/oradwell/covers-validator/pull/48.